server: fixes
